fix(quic): correct UDP_SEGMENT probe optlen so GSO actually engages
supportsGso() probed UDP_SEGMENT with a 2-byte u16, but the setsockopt takes
a 4-byte int — it returned EINVAL on every Linux kernel (even those that
fully support GSO), so quic_gso was always false and the GSO send path was
dead code: a 200 KB h3 response did ~170 sendto calls, zero sendmsg
(confirmed via strace on a kernel-6.17 box). Probe with c_int. The per-send
cmsg payload stays u16 (correct for the UDP_SEGMENT control message).

+++ src/runtime/net.zig
@@ -404,13 +404,17 @@ pub fn supportsGso() bool {
     const fd = std.posix.system.socket(std.posix.AF.INET, std.posix.SOCK.DGRAM, std.posix.IPPROTO.UDP);
     if (fd < 0) return false;
     defer clock.closeFd(fd);
-    var val: u16 = 1280;
+    // The UDP_SEGMENT *setsockopt* takes an `int` (4 bytes). Probing with a
+    // 2-byte value returns EINVAL even on kernels that fully support GSO,
+    // which silently disabled the GSO send path entirely. (The per-message
+    // cmsg payload, by contrast, is a u16 — see sendGsoLinux.)
+    var val: c_int = 1280;
     const rc = std.posix.system.setsockopt(
         fd,
         SOL_UDP,
         UDP_SEGMENT,
         @ptrCast(&val),
-        @sizeOf(u16),
+        @sizeOf(c_int),
     );
     return rc == 0;
 }
